Update H.ahkBoxGUI081021.py
figured out  the scope issues made the df value a global value so every portion has access to if they want.
Variables should almost never be global but for now this is the best option for making it available to other portions of code 
with out having to create a whole separate function and calling that function twice, by the graph and download button, just to get the array values. 
this was a lot easier if it causes issues somehow later on than it should be changed

I also changed the t4 value to t just because it caused an issue with graph.