"some what" finalized version of arduino test code
this version has shown to work as intended. it can randomly decide to send data and decide when to not send data. it can send random values and random number of bytes
the only thing left is to test that it can communicate when there are other i2c devices connected on the same line.