created Card Library python module
we may not need an entirely separate module for handing card information and formatting but currently it might be a good idea as to keep things a bit more organized and less confusing and cluttered. that way the h.ahkBox.py can focus on its baseline function and communication. All while this module can be used to interpret the byte data from given cards and translate the bytes into a format that can be displayed on the GUI and placed into a file to be downloaded and viewed. 

note: (may or may not be needed) we may also need to have a seprate file that has the raw byte information saved as well so that incase there is weird error we can see bit by bit where that error could be occuring-