fixing graph and file exporting errors
was trying to add a way to update and information to a file that can be downloaded from a button however in the process I some how broke the graph. when i get the graph working again i will update this patch and when i get everything else working without errors i will commit it to the main branch file.