new daily files
made it so that the library will create a new file everyday. this current way was based off the easiest way I could think of. it will simply name itself the day month and year and when the day changes it will create a new file with that name. the next step I still need to figure out is a way so that the data can be appended to the same file throughout the day as it runs.