updated channels
only channel one had the proper formatting. so i changed it so all the other channels had the same kind of formatting.  I copy and pasted channel ones formatting for each channel so if there are any odd errors may need to double check that I didn't miss any variable names and that i have the right values int the if statements.