testing just the h.ahk  version
this is for testing the h.ahk version. if you want to test out the card library you will need to make sure the arduino is sending very specific values and values sizes other wise the results will not be accurate. 
as an example for what the arduino can send if you want to test it:
(0x33, 0x88,    0x01, 0x02, 0xa3, 0x04, 0x05,)
add this to data portion fo the arduino code instead of the function that creates adn random array size and random values in that arrary.